requirements.txt
flask
tensorflow-cpu
opencv-contrib-python-headless
numpy
gunicorn